% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galah_download.R
\name{download_ala_obs}
\alias{download_ala_obs}
\title{Download observations for Infinity List}
\usage{
download_ala_obs(
  taxa,
  year_range = c(1923, 2023),
  save_raw_data = FALSE,
  output_dir = system.file(package = "infinitylists", "data/")
)
}
\arguments{
\item{taxa}{}

\item{year_range}{}

\item{save_raw_data}{}

\item{output_dir}{}
}
\description{
Download observations for Infinity List
}
