% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_download.R
\name{query_gbif_global}
\alias{query_gbif_global}
\title{Generate galah query for Global GBIF Node}
\usage{
query_gbif_global(taxon, min_year, max_year, country_code = NULL)
}
\arguments{
\item{taxon}{character, genus/family/kingdom}

\item{min_year}{numeric, year cut off for query, only records where year >= min_year will be included}

\item{country_code}{character, code for country}
}
\description{
Generate galah query for Global GBIF Node
}
